+++ pkg/streamingpromql/planning/plan.go
@@ -239,6 +239,7 @@ type OperatorParameters struct {
 	Annotations              *annotations.Annotations
 	QueryStats               *types.QueryStats
 	EagerLoadSelectors       bool
+	CacheDisabled            bool
 	QueryParameters          *QueryParameters
 	Logger                   log.Logger
 }